Conflicts:
	sql/base/world_database.sql
	src/server/game/AI/CoreAI/CombatAI.cpp
	src/server/game/AI/CoreAI/CombatAI.h
	src/server/game/AI/CoreAI/PassiveAI.h
	src/server/game/AI/SmartScripts/SmartScript.cpp
	src/server/game/Battlegrounds/ArenaTeam.cpp
	src/server/game/Battlegrounds/Battleground.cpp
	src/server/game/Battlegrounds/BattlegroundMgr.cpp
	src/server/game/Battlegrounds/BattlegroundQueue.cpp
	src/server/game/Battlegrounds/Zones/BattlegroundAV.h
	src/server/game/Battlegrounds/Zones/BattlegroundEY.h
	src/server/game/Battlegrounds/Zones/BattlegroundWS.cpp
	src/server/game/Battlegrounds/Zones/BattlegroundWS.h
	src/server/game/Chat/Commands/Level2.cpp
	src/server/game/Chat/Commands/Level3.cpp
	src/server/game/Entities/Creature/Creature.cpp
	src/server/game/Entities/Creature/CreatureGroups.h
	src/server/game/Entities/Creature/GossipDef.cpp
	src/server/game/Entities/Creature/GossipDef.h
	src/server/game/Entities/GameObject/GameObject.cpp
	src/server/game/Entities/Item/Item.h
	src/server/game/Entities/Object/Object.cpp
	src/server/game/Entities/Object/ObjectPosSelector.cpp
	src/server/game/Entities/Pet/Pet.cpp
	src/server/game/Entities/Pet/Pet.h
	src/server/game/Entities/Player/Player.cpp
	src/server/game/Entities/Player/Player.h
	src/server/game/Entities/Player/SocialMgr.h
	src/server/game/Entities/Transport/Transport.cpp
	src/server/game/Entities/Unit/Unit.cpp
	src/server/game/Entities/Unit/Unit.h
	src/server/game/Entities/Vehicle/Vehicle.cpp
	src/server/game/Globals/ObjectMgr.cpp
	src/server/game/Grids/Notifiers/GridNotifiers.cpp
	src/server/game/Maps/Map.cpp
	src/server/game/Miscellaneous/SharedDefines.h
	src/server/game/OutdoorPvP/OutdoorPvP.cpp
	src/server/game/OutdoorPvP/OutdoorPvP.h
	src/server/game/OutdoorPvP/OutdoorPvPMgr.cpp
	src/server/game/OutdoorPvP/OutdoorPvPMgr.h
	src/server/game/Scripting/MapScripts.cpp
	src/server/game/Scripting/ScriptLoader.cpp
	src/server/game/Scripting/ScriptMgr.cpp
	src/server/game/Scripting/ScriptMgr.h
	src/server/game/Server/Protocol/Handlers/GroupHandler.cpp
	src/server/game/Server/Protocol/Handlers/MiscHandler.cpp
	src/server/game/Server/Protocol/Handlers/QuestHandler.cpp
	src/server/game/Server/WorldSession.h
	src/server/game/Spells/Auras/SpellAuraEffects.cpp
	src/server/game/Spells/Auras/SpellAuraEffects.h
	src/server/game/Spells/Auras/SpellAuras.cpp
	src/server/game/Spells/Auras/SpellAuras.h
	src/server/game/Spells/Spell.cpp
	src/server/game/Spells/Spell.h
	src/server/game/Spells/SpellEffects.cpp
	src/server/game/Spells/SpellMgr.cpp
	src/server/game/Spells/SpellMgr.h
	src/server/game/Spells/SpellScript.h
	src/server/game/World/World.cpp
	src/server/game/World/World.h
	src/server/scripts/Kalimdor/RuinsOfAhnQiraj/boss_ayamiss.cpp
	src/server/scripts/Kalimdor/RuinsOfAhnQiraj/boss_buru.cpp
	src/server/scripts/Kalimdor/RuinsOfAhnQiraj/boss_kurinnaxx.cpp
	src/server/scripts/Kalimdor/RuinsOfAhnQiraj/boss_moam.cpp
	src/server/scripts/Kalimdor/RuinsOfAhnQiraj/boss_ossirian.cpp
	src/server/scripts/Kalimdor/RuinsOfAhnQiraj/boss_rajaxx.cpp
	src/server/scripts/Kalimdor/TempleOfAhnQiraj/boss_bug_trio.cpp
	src/server/scripts/Kalimdor/TempleOfAhnQiraj/boss_fankriss.cpp
	src/server/scripts/Kalimdor/TempleOfAhnQiraj/boss_sartura.cpp
	src/server/scripts/Spells/spell_shaman.cpp

+++ src/server/game/Spells/Spell.h
@@ -169,7 +169,7 @@ class SpellCastTargets
         void setTargetMask(uint32 newMask) { m_targetMask = newMask; }
 
         uint64 getUnitTargetGUID() const { return m_unitTargetGUID; }
-        Unit* getUnitTarget() const { return m_unitTarget; }
+        Unit *getUnitTarget() const { return m_unitTarget; }
         void setUnitTarget(Unit* target);
         void setSrc(float x, float y, float z);
         void setSrc(Position &pos);
@@ -225,7 +225,7 @@ class SpellCastTargets
     private:
         uint32 m_targetMask;
         // objects (can be used at spell creating and after Update at casting
-        Unit* m_unitTarget;
+        Unit *m_unitTarget;
         GameObject *m_GOTarget;
         Item *m_itemTarget;
 
@@ -241,7 +241,7 @@ struct SpellValue
 {
     explicit SpellValue(SpellEntry const* proto)
     {
-        for (uint8 i = 0; i < MAX_SPELL_EFFECTS; ++i)
+        for (uint32 i = 0; i < MAX_SPELL_EFFECTS; ++i)
             EffectBasePoints[i] = proto->EffectBasePoints[i];
         MaxAffectedTargets = proto->MaxAffectedTargets;
         RadiusMod = 1.0f;
@@ -285,7 +285,7 @@ enum SpellTargets
 class Spell
 {
     friend struct Trinity::SpellNotifierCreatureAndPlayer;
-    friend void Unit::SetCurrentCastedSpell(Spell * pSpell);
+    friend void Unit::SetCurrentCastedSpell(Spell* pSpell);
     friend class SpellScript;
     public:
 
@@ -416,9 +416,9 @@ class Spell
         void EffectCastButtons(SpellEffIndex effIndex);
         void EffectRechargeManaGem(SpellEffIndex effIndex);
 
-        typedef std::set<Aura* > UsedSpellMods;
+        typedef std::set<Aura *> UsedSpellMods;
 
-        Spell(Unit* Caster, SpellEntry const* info, bool triggered, uint64 originalCasterGUID = 0, bool skipCheck = false);
+        Spell(Unit* Caster, SpellEntry const *info, bool triggered, uint64 originalCasterGUID = 0, bool skipCheck = false);
         ~Spell();
 
         void prepare(SpellCastTargets const* targets, AuraEffect const* triggeredByAura = NULL);
@@ -458,8 +458,8 @@ class Spell
         void setState(uint32 state) { m_spellState = state; }
 
         void DoCreateItem(uint32 i, uint32 itemtype);
-        void WriteSpellGoTargets(WorldPacket* data);
-        void WriteAmmoToPacket(WorldPacket* data);
+        void WriteSpellGoTargets(WorldPacket * data);
+        void WriteAmmoToPacket(WorldPacket * data);
 
         void SelectSpellTargets();
         void SelectEffectTargets(uint32 i, uint32 cur);
@@ -494,7 +494,7 @@ class Spell
         void SendResurrectRequest(Player* target);
         void SendPlaySpellVisual(uint32 SpellID);
 
-        void HandleEffects(Unit* pUnitTarget, Item *pItemTarget, GameObject *pGOTarget, uint32 i);
+        void HandleEffects(Unit *pUnitTarget, Item *pItemTarget, GameObject *pGOTarget, uint32 i);
         void HandleThreatSpells(uint32 spellId);
 
         const SpellEntry * const m_spellInfo;
@@ -599,7 +599,7 @@ class Spell
         GameObject* gameObjTarget;
         int32 damage;
         // used in effects handlers
-        Aura* m_spellAura;
+        Aura * m_spellAura;
 
         // this is set in Spell Hit, but used in Apply Aura handler
         DiminishingLevels m_diminishLevel;
@@ -662,8 +662,8 @@ class Spell
         void AddGOTarget(uint64 goGUID, uint32 effIndex);
         void AddItemTarget(Item* target, uint32 effIndex);
         void DoAllEffectOnTarget(TargetInfo *target);
-        SpellMissInfo DoSpellHitOnUnit(Unit* unit, uint32 effectMask, bool scaleAura);
-        void DoTriggersOnSpellHit(Unit* unit);
+        SpellMissInfo DoSpellHitOnUnit(Unit *unit, uint32 effectMask, bool scaleAura);
+        void DoTriggersOnSpellHit(Unit *unit);
         void DoAllEffectOnTarget(GOTargetInfo *target);
         void DoAllEffectOnTarget(ItemTargetInfo *target);
         bool UpdateChanneledTargetList();
@@ -675,7 +675,7 @@ class Spell
         bool IsValidSingleTargetSpell(Unit const* target) const;
         bool IsValidDeadOrAliveTarget(Unit const* target) const;
         void CalculateDamageDoneForAllTargets();
-        int32 CalculateDamageDone(Unit* unit, const uint32 effectMask, float *multiplier);
+        int32 CalculateDamageDone(Unit *unit, const uint32 effectMask, float *multiplier);
         void SpellDamageSchoolDmg(SpellEffIndex effIndex);
         void SpellDamageWeaponDmg(SpellEffIndex effIndex);
         void SpellDamageHeal(SpellEffIndex effIndex);
@@ -701,7 +701,7 @@ class Spell
 
         // effect helpers
         void GetSummonPosition(uint32 i, Position &pos, float radius = 0.0f, uint32 count = 0);
-        void SummonGuardian(uint32 i, uint32 entry, SummonPropertiesEntry const* properties);
+        void SummonGuardian(uint32 i, uint32 entry, SummonPropertiesEntry const *properties);
         void CalculateJumpSpeeds(uint8 i, float dist, float & speedxy, float & speedz);
 
         SpellCastResult CanOpenLock(uint32 effIndex, uint32 lockid, SkillType& skillid, int32& reqSkillValue, int32& skillValue);
@@ -751,7 +751,7 @@ namespace Trinity
         bool i_requireDeadTarget;
         SpellEntry const* i_spellProto;
 
-        SpellNotifierCreatureAndPlayer(Unit* source, std::list<Unit*> &data, float radius, SpellNotifyPushType type,
+        SpellNotifierCreatureAndPlayer(Unit *source, std::list<Unit*> &data, float radius, SpellNotifyPushType type,
             SpellTargets TargetType = SPELL_TARGETS_ENEMY, const Position *pos = NULL, uint32 entry = 0, SpellEntry const* spellProto = NULL)
             : i_data(&data), i_push_type(type), i_radius(radius), i_TargetType(TargetType),
             i_source(source), i_entry(entry), i_pos(pos), i_spellProto(spellProto)
